refactor: 패키지 세분화
