--- src/main/kotlin/racingcar/RacingTrial.kt
@@ -1,3 +1,0 @@
-package racingcar
-
-class RacingTrial(val locations: List<CarLocation>)
