+++ src/main/kotlin/racingcar/racing/RacingResults.kt
@@ -1,3 +1,3 @@
-package racingcar
+package racingcar.racing
 
 class RacingResults(val trials: List<RacingTrial>)
