+++ src/main/kotlin/racingcar/racing/RacingTrial.kt
@@ -1,0 +1,5 @@
+package racingcar.racing
+
+import racingcar.racing.car.CarLocation
+
+class RacingTrial(val locations: List<CarLocation>)
