+++ src/main/kotlin/racingcar/racing/car/Car.kt
@@ -1,4 +1,6 @@
-package racingcar
+package racingcar.racing.car
+
+import racingcar.racing.car.engine.Engine
 
 class Car(
     private val engine: Engine
