+++ src/main/kotlin/racingcar/racing/car/CarLocation.kt
@@ -1,4 +1,6 @@
-package racingcar
+package racingcar.racing.car
+
+import racingcar.racing.car.engine.Torque
 
 data class CarLocation(val value: Int) {
     fun moveBy(torque: Torque): CarLocation {
