+++ src/main/kotlin/racingcar/racing/car/engine/Engine.kt
@@ -1,4 +1,4 @@
-package racingcar
+package racingcar.racing.car.engine
 
 interface Engine {
     fun run(): Torque
