+++ src/main/kotlin/racingcar/racing/car/engine/ForceGenerator.kt
@@ -1,4 +1,4 @@
-package racingcar
+package racingcar.racing.car.engine
 
 interface ForceGenerator {
     fun generate(): Force
