+++ src/main/kotlin/racingcar/racing/car/engine/RacingEngine.kt
@@ -1,4 +1,4 @@
-package racingcar
+package racingcar.racing.car.engine
 
 class RacingEngine(
     private val threshold: Force = DEFAULT_THRESHOLD,
