+++ src/main/kotlin/racingcar/racing/car/engine/Torque.kt
@@ -1,4 +1,4 @@
-package racingcar
+package racingcar.racing.car.engine
 
 /**
  * [Engine]이 발생시키는 돌림힘.
