+++ src/main/kotlin/racingcar/view/InputView.kt
@@ -1,4 +1,4 @@
-package racingcar
+package racingcar.view
 
 class InputView(
     private val carNumberQuestion: String = "자동차 대수는 몇 대인가요?",
