+++ src/test/kotlin/racingcar/racing/car/CarLocationTest.kt
@@ -1,10 +1,11 @@
-package racingcar
+package racingcar.racing.car
 
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.CsvSource
 import org.junit.jupiter.params.provider.ValueSource
+import racingcar.racing.car.engine.Torque
 
 internal class CarLocationTest {
     @Test
